+++ firebase-config/src/main/java/com/google/firebase/remoteconfig/internal/ConfigRealtimeHttpClient.java
@@ -1,6 +1,7 @@
 package com.google.firebase.remoteconfig.internal;
 
-import com.google.firebase.remoteconfig.FirebaseRemoteConfig;
+import com.google.firebase.remoteconfig.ConfigUpdateListener;
+import com.google.firebase.remoteconfig.ConfigUpdateListenerRegistration;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -11,7 +12,7 @@ import java.util.Map;
  */
 public class ConfigRealtimeHttpClient {
 
-  private final Map<Integer, FirebaseRemoteConfig.ConfigUpdateListener> listeners;
+  private final Map<Integer, ConfigUpdateListener> listeners;
   private int listenerCount;
 
   public ConfigRealtimeHttpClient() {
@@ -25,11 +26,11 @@ public class ConfigRealtimeHttpClient {
   // Pauses Http stream listening
   private void pauseRealtimeStream() {}
 
-  public FirebaseRemoteConfig.ConfigUpdateListenerRegistration addRealtimeConfigUpdateListener(
-      FirebaseRemoteConfig.ConfigUpdateListener configUpdateListener) {
+  public ConfigUpdateListenerRegistration addRealtimeConfigUpdateListener(
+      ConfigUpdateListener configUpdateListener) {
     listeners.put(listenerCount, configUpdateListener);
     beginRealtimeStream();
-    return new FirebaseRemoteConfig.ConfigUpdateListenerRegistration(this, listenerCount++);
+    return new ConfigUpdateListenerRegistration(this, listenerCount++);
   }
 
   public void removeRealtimeConfigUpdateListener(int listenerKey) {
